PythonLibraries/requirements.txt
asn1crypto==0.24.0
azure-common==1.1.23
azure-cosmosdb-nspkg==2.0.2
azure-cosmosdb-table==1.0.5
azure-functions==1.0.0b3
azure-functions-worker==1.0.0b4
azure-nspkg==3.0.2
azure-storage-blob==1.5.0
azure-storage-common==1.4.2
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.6.1
entrypoints==0.3
flake8==3.7.8
grpcio==1.14.2
grpcio-tools==1.14.2
idna==2.8
joblib==0.13.2
mccabe==0.6.1
numpy==1.16.2
pandas==0.24.1
protobuf==3.7.0
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1
requests==2.22.0
ruamel.appconfig==0.5.4
ruamel.std.argparse==0.8.1
scikit-learn==0.21.3
scipy==1.2.1
six==1.12.0
sklearn==0.0
urllib3==1.25.3
yaml-1.3==0.1.0
